fix: 修复静态扫描确认的 69 处安全缺陷
对静态代码扫描报告逐条核实源码，确认 69 条确实存在缺陷（30 条真实问题 +
39 条部分成立），覆盖资源泄漏、数组越界、整数运算溢出反转/溢出、空指针、
无限循环、数据竞争等类型；其余 366 条经核实为误报，未做改动。

异常路径行为调整：
- ExecLocalStateHandler 由无限重试改为上限 300 次（约 10 分钟）
- RmObmmExecutor::ObmmUnExport 单个 memId 失败后继续释放剩余 memId
- 预上线应答循环由 while(true) 改为固定 5 次
- ShareBorrowRpcObjCheck 对 IMPORT 对象同样拒绝空 exportNumaInfos

验证：bash build.sh ut 全量编译 + 5053 个用例全部通过，无编译错误。
未新增测试用例；2 个既有 UT 因源码行为变化补充了 mock 桩。

+++ src/adapter_plugins/mti/out_of_band/message/ubse_ictrl_q_req_msg.cpp
@@ -41,12 +41,12 @@ static void SetServiceTypeToBuf(uint8_t serviceType, CtrlQReqMessage& reqMsg)
     reqMsg.blocks.front().head.serviceType = serviceType;
 }
 
-static void SetBBNumToBuf(uint8_t bbNum, CtrlQReqMessage& reqMsg)
+static void SetBBNumToBuf(uint32_t bbNum, CtrlQReqMessage& reqMsg)
 {
-    reqMsg.blocks.front().head.bbNum = bbNum;
+    reqMsg.blocks.front().head.bbNum = static_cast<uint8_t>(bbNum);
 }
 
-ICtrlQReqMsg::ICtrlQReqMsg(uint8_t opCode, uint8_t bbNum)
+ICtrlQReqMsg::ICtrlQReqMsg(uint8_t opCode, uint32_t bbNum)
 {
     reqMsg_ = CtrlQReqMessage(bbNum);
     SetBBNumToBuf(bbNum, reqMsg_);
@@ -89,7 +89,7 @@ void ICtrlQReqMsg::SetServiceType(uint8_t serviceType)
     SetServiceTypeToBuf(serviceType, reqMsg_);
 }
 
-void ICtrlQReqMsg::SetBBNum(uint8_t bbNum)
+void ICtrlQReqMsg::SetBBNum(uint32_t bbNum)
 {
     SetBBNumToBuf(bbNum, reqMsg_);
 }
